c++ update, clang-format (#49)
* c++ update, clang-format, sketch new api

* array

* c++ COBS_ENCODE_MAX plus tests

* fix windows

* more c++ cleanups

* doctest upgrade

* take out incremental decode

+++ tests/test_cobs_decode.cc
@@ -4,23 +4,17 @@
 
 TEST_CASE("Decoding validation") {
   unsigned char dec[32];
-  unsigned dec_len;
+  size_t dec_len;
 
   SUBCASE("Invalid payload: jump past end") {
-    byte_vec_t enc{3, 0};
-    REQUIRE(cobs_decode(enc.data(),
-                        unsigned(enc.size()),
-                        dec,
-                        sizeof(dec),
-                        &dec_len) == COBS_RET_ERR_BAD_PAYLOAD);
+    byte_vec_t enc{ 3, 0 };
+    REQUIRE(cobs_decode(enc.data(), enc.size(), dec, sizeof(dec), &dec_len) ==
+            COBS_RET_ERR_BAD_PAYLOAD);
   }
 
   SUBCASE("Invalid payload: jump over internal zeroes") {
-    byte_vec_t enc{5, 1, 0, 0, 1, 0};
-    REQUIRE(cobs_decode(enc.data(),
-                        unsigned(enc.size()),
-                        dec,
-                        sizeof(dec),
-                        &dec_len) == COBS_RET_ERR_BAD_PAYLOAD);
+    byte_vec_t enc{ 5, 1, 0, 0, 1, 0 };
+    REQUIRE(cobs_decode(enc.data(), enc.size(), dec, sizeof(dec), &dec_len) ==
+            COBS_RET_ERR_BAD_PAYLOAD);
   }
 }
